Use Kotest matchers
